AppKit works end to end

+++ src/Frameworks/CoreGraphics/CGContext.h
@@ -159,7 +159,7 @@ COREGRAPHICS_EXPORT void CGContextClipToRects(CGContextRef context, const CGRect
 COREGRAPHICS_EXPORT void CGContextSetStrokeColorSpace(CGContextRef context, CGColorSpaceRef colorSpace);
 COREGRAPHICS_EXPORT void CGContextSetFillColorSpace(CGContextRef context, CGColorSpaceRef colorSpace);
 
-COREGRAPHICS_EXPORT void CGContextSetStrokeColor(CGContextRef context, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetStrokeColor(CGContextRef context, const CGFloat *components);
 COREGRAPHICS_EXPORT void CGContextSetStrokeColorWithColor(CGContextRef context, CGColorRef color);
 COREGRAPHICS_EXPORT void CGContextSetGrayStrokeColor(CGContextRef context, float gray, float alpha);
 COREGRAPHICS_EXPORT void CGContextSetRGBStrokeColor(CGContextRef context, float r, float g, float b, float alpha);
@@ -167,7 +167,7 @@ COREGRAPHICS_EXPORT void CGContextSetCMYKStrokeColor(CGContextRef context, float
 COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBStrokeColor(CGContextRef context, float red, float green, float blue, float alpha);
 COREGRAPHICS_EXPORT void CGContextSetCalibratedGrayStrokeColor(CGContextRef context, float gray, float alpha);
 
-COREGRAPHICS_EXPORT void CGContextSetFillColor(CGContextRef context, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetFillColor(CGContextRef context, const CGFloat *components);
 COREGRAPHICS_EXPORT void CGContextSetFillColorWithColor(CGContextRef context, CGColorRef color);
 COREGRAPHICS_EXPORT void CGContextSetGrayFillColor(CGContextRef context, float gray, float alpha);
 COREGRAPHICS_EXPORT void CGContextSetRGBFillColor(CGContextRef context, float r, float g, float b, float alpha);
@@ -178,8 +178,8 @@ COREGRAPHICS_EXPORT void CGContextSetCalibratedRGBFillColor(CGContextRef context
 COREGRAPHICS_EXPORT void CGContextSetAlpha(CGContextRef context, float alpha);
 
 COREGRAPHICS_EXPORT void CGContextSetPatternPhase(CGContextRef context, CGSize phase);
-COREGRAPHICS_EXPORT void CGContextSetStrokePattern(CGContextRef context, CGPatternRef pattern, const float *components);
-COREGRAPHICS_EXPORT void CGContextSetFillPattern(CGContextRef context, CGPatternRef pattern, const float *components);
+COREGRAPHICS_EXPORT void CGContextSetStrokePattern(CGContextRef context, CGPatternRef pattern, const CGFloat *components);
+COREGRAPHICS_EXPORT void CGContextSetFillPattern(CGContextRef context, CGPatternRef pattern, const CGFloat *components);
 
 COREGRAPHICS_EXPORT void CGContextSetTextMatrix(CGContextRef context, CGAffineTransform matrix);
