+++ src/Frameworks/Onyx2D/O2Surface.m
@@ -730,7 +730,7 @@ void O2GaussianKernelRelease(O2GaussianKernelRef kernel) {
 
 static O2argb32f argbFromColor(O2ColorRef color){   
    size_t    count=O2ColorGetNumberOfComponents(color);
-   const float *components=O2ColorGetComponents(color);
+   const CGFloat *components=O2ColorGetComponents(color);
 
    if(count==2)
     return O2argb32fInit(components[0],components[0],components[0],components[1]);
